Monolithic/Service/requirements.txt
flask
pydub
openai-whisper
ffmpeg-python
torch
tqdm
numpy
minio
